Introduce Drush (and update a bunch of dependencies 🙈) (#136)
* Update dependencies and install drupal/dev and Drush... read on...

I know what you're thinking, this should have been split into a couple of commits. This was stymied by my lack of
composer knowledge and a couple of challenges:

- From https://github.com/drupal-composer/drupal-project/blob/9.x/composer.json, we need to add  before we can get  to work
- Installing either of these fails because symfony is locked to 5.1.7. I found a recommendation to split the require and update steps, which worked
here: https://www.drupal.org/project/drupalmoduleupgrader/issues/3072073.

So we ended up with the following commands:

composer require --dev drupal/core-dev ^8.9 --no-update
composer require drush/drush--no-update
composer update

Which is the entirety of this commit. I am sorry to my future self and any reviewers that this includes a bunch of dependency upgrades
in the same place, but I lost 30 minutes trying to unpick it and didn't get anywhere. Brute Force for the win.

* Trivial default settings comment changes, introduced by composer on update